// src/data/pokemon/types.ts
export const TYPES: TypeEffectiveness[] = [
  {
    "id": 13,
    "name": "bug",
    "damageTaken": {
      "bug": 0,
      "dark": 0,
      "dragon": 0,
      "electric": 0,
      "fairy": 0,
      "fighting": 2,
      "fire": 1,
      "flying": 1,
      "ghost": 0,
      "grass": 2,
      "ground": 2,
      "ice": 0,
      "normal": 0,
      "poison": 0,
      "psychic": 0,
      "rock": 1,
      "steel": 0,
      "water": 0
    }
  },
// ...
];